feat: перенесла валидацию в бизнес-логику фильмов
