+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,14 +1,12 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
-import ru.yandex.practicum.filmorate.throwable.ValidationException;
 
 import javax.validation.Valid;
-import java.time.LocalDate;
-import java.time.Month;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,9 +14,9 @@ import java.util.List;
 @RestController
 @RequestMapping("/films")
 public class FilmController {
-    private static final LocalDate AFTER_RELEASE_DATE = LocalDate.of(1895, Month.DECEMBER, 28);
     private final FilmService filmService;
 
+    @Autowired
     public FilmController(FilmService filmService) {
         this.filmService = filmService;
     }
@@ -31,14 +29,12 @@ public class FilmController {
 
     @PostMapping
     public Film create(@Valid @RequestBody Film film) {
-        validateFilm(film);
         filmService.createFilm(film);
         return film;
     }
 
     @PutMapping
     public Film update(@Valid @RequestBody Film film) {
-        validateFilm(film);
         filmService.updateFilm(film);
         return film;
     }
@@ -65,14 +61,5 @@ public class FilmController {
     @DeleteMapping("/{id}/like/{userId}")
     public void removeLike(@PathVariable("id") int id, @PathVariable("userId") int userId) {
         filmService.removeLike(id, userId);
-    }
-
-    private void validateFilm(Film film) {
-        if (film.getReleaseDate().isBefore(AFTER_RELEASE_DATE)) {
-            log.error("Не прошёл валидацию.");
-            throw new ValidationException("Фильм должен быть не раньше " + AFTER_RELEASE_DATE.getDayOfMonth()
-                    + " " + AFTER_RELEASE_DATE.getMonth() + " " + AFTER_RELEASE_DATE.getYear());
-        }
-        log.debug(film.toString());
     }
 }
